pyproject.toml
[tool.poetry]
name = "db-performance-testing"
version = "0.1.0"
description = "This repo contains code for testing the performance of databases using locust and unicorn, It also contains code for testing the performance of databases "
authors = ["ekharshavardhan851@gmail.com"]
readme = "README.md"
packages = [
    {include = "api", from= "app"},
    {include = "performance", from= "app"}
]
repository = "https://github.com/sh-harshavardhan/db-performance-testing.git"


[tool.poetry.dependencies]
python = ">=3.13,<4.0.0"
pydantic = ">=2.12.5,<3.0.0"
pydantic-settings = ">=2.12.0,<3.0.0"
duckdb = ">=1.4.4,<2.0.0"
jinja2 = ">=3.1.6,<4.0.0"
pyarrow = ">=23.0.0,<24.0.0"
faker = "^40.4.0"
typer = "^0.21.1"
gitpython = "^3.1.46"


[tool.poetry.scripts]
db-performance = 'performance.src.cli:app'

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.5.1"
sphinxcontrib-typer = "^0.7.2"

[tool.poetry.group.iceberg]
optional = true

[tool.poetry.group.iceberg.dependencies]
pyiceberg = {extras = ["all"], version=">=0.10.0,<0.11.0"}


[tool.poetry.group.api]
optional = true

[tool.poetry.group.api.dependencies]
fastapi = "^0.128.3"
sqlalchemy = "^2.0.46"
locust = "^2.43.2"
locust-plugins = "^5.0.0"
unicorn = "^2.1.4"

[tool.poetry.group.snowflake]
optional = true


[tool.poetry.group.snowflake.dependencies]
snowflake-connector-python = "^4.2.0"
snowflake-sqlalchemy = "^1.8.2"

[tool.poetry.group.postgres]
optional = true

[tool.poetry.group.postgres.dependencies]
psycopg2-binary = "^2.9.11"
psycopg2 = "^2.9.11"


[tool.poetry.group.databricks]
optional = true

[tool.poetry.group.databricks.dependencies]
databricks-sqlalchemy = "^2.0.8"
databricks-sql-connector = "^4.2.4"


[tool.ruff]

fix = true
target-version = "py38"
exclude = [
    ".git",
    ".mypy_cache",
    ".pyenv",
    ".pytest_cache",
    ".ruff_cache",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

include = [
    "pyproject.toml",
    "app/**/*.py"
]


line-length = 120
indent-width = 4


[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = true

[tool.ruff.lint]
extend-select = [
  "UP",  # pyupgrade
  "D",   # pydocstyle
]

ignore = ["D415", "D205"]

[tool.ruff.lint.pydocstyle]
convention = "google"
